requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
attrs==26.1.0
autopep8==2.3.2
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
frozenlist==1.8.0
googleapis-common-protos==1.73.1
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
multidict==6.7.1
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
outcome==1.3.0.post0
packaging==25.0
propcache==0.4.1
protobuf==6.33.6
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.33.0
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==1.0.0
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
websocket-client==1.9.0
wsproto==1.3.2
xai-sdk==1.11.0
yarl==1.23.0
zipp==3.23.0
